Reverting default configurations from condDB v1 to v2
