+++ Configuration/StandardSequences/python/FrontierConditions_GlobalTag_condDBv1_cfi.py
@@ -2,11 +2,10 @@ import FWCore.ParameterSet.Config as cms
 
 from CondCore.DBCommon.CondDBSetup_cfi import *
 
-print '# Conditions read from  CMS_CONDITIONS  via FrontierProd '
-
 GlobalTag = cms.ESSource("PoolDBESSource",
     CondDBSetup,
-    connect = cms.string('frontier://FrontierProd/CMS_CONDITIONS'),
-    globaltag = cms.string('UNSPECIFIED'),
+    connect = cms.string('frontier://FrontierProd/CMS_COND_31X_GLOBALTAG'),
+    globaltag = cms.string('UNSPECIFIED::All'),
     toGet = cms.VPSet( ),   # hook to override or add single payloads
+    BlobStreamerName = cms.untracked.string('TBufferBlobStreamingService')
 )
